+++ system-tests/tests/smoke/cre/evm/evmread/contracts/MessageEmitter.bin
