+++ system-tests/tests/smoke/cre/evm/evmread/go.mod
@@ -1,4 +1,4 @@
-module github.com/smartcontractkit/chainlink/system-tests/tests/smoke/cre/evmread
+module github.com/smartcontractkit/chainlink/system-tests/tests/smoke/cre/evm/evmread
 
 go 1.24.5
 
@@ -6,6 +6,7 @@ require (
 	github.com/ethereum/go-ethereum v1.16.3
 	github.com/google/go-cmp v0.7.0
 	github.com/smartcontractkit/chainlink-protos/cre/go v0.0.0-20250829155125-f4655b0b4605
+	github.com/smartcontractkit/chainlink/system-tests/tests/smoke/cre/evmread v0.0.0-20250917232237-c4ecf802c6f8
 	github.com/smartcontractkit/cre-sdk-go v0.6.0
 	github.com/smartcontractkit/cre-sdk-go/capabilities/blockchain/evm v0.6.0
 	github.com/smartcontractkit/cre-sdk-go/capabilities/scheduler/cron v0.6.0
