+++ system-tests/tests/smoke/cre/evm/evmread/main.go
